Created util functions file to avoid code repetition
