+++ armadillo_cli.py
@@ -1,8 +1,13 @@
 import typer
+import os
 
+import utils
 import armadillo_cli_add
 import armadillo_cli_list
 import armadillo_cli_delete
+
+if not os.path.exists(utils.armadillo_dir_path):
+    os.mkdir(utils.armadillo_dir_path)
 
 app = typer.Typer()
 app.add_typer(armadillo_cli_add.app, name="add", help="Add new policy      (policy|policy-dns|policy-dpi)")
